refactor(webview): remove Webview namespace
